+++ lib/redis_queued_locks/acquier/acquire_lock.rb
@@ -72,7 +72,8 @@ module RedisQueuedLocks::Acquier::AcquireLock
     #   - Metadata can not contain reserved lock data keys;
     # @option logger [::Logger,#debug]
     #   - Logger object used from the configuration layer (see config[:logger]);
-    #   - See RedisQueuedLocks::Logging::VoidLogger for example;
+    #   - See `RedisQueuedLocks::Logging::VoidLogger` for example;
+    #   - Supports `SemanticLogger::Logger` (see "semantic_logger" gem)
     # @option log_lock_try [Boolean]
     #   - should be logged the each try of lock acquiring (a lot of logs can be generated depending
     #     on your retry configurations);