+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [Unreleased]
 ### Added
 - **Major Feature**: support for **Reentrant Locks** (super strong and diffcult feature):
-- The result of lock obtaining now includes `:process` key that shows a type of logical process that obtains a lock
+- The result of lock obtaining now includes `:process` key that shows the type of logical process that obtains the lock
   (`:lock_obtaining`, `:extendable_conflict_work_through`, `:conflict_work_through`);
 - Added reentrant lock details to `RedisQueuedLocks::Client#lock_info` and `RedisQueuedLocks::Client#locks` method results;
 - Documentation updates;